extract jsonValidator and WrappedJsonSchema
