+++ components/ledger/ledger-persistence/build.gradle
@@ -39,6 +39,7 @@ dependencies {
     implementation project(':libs:messaging:messaging')
     implementation project(':libs:metrics')
     implementation project(':libs:serialization:json-validator')
+    implementation project(':libs:serialization:json-validator-common')
     implementation project(':libs:utilities')
     implementation project(":libs:serialization:serialization-avro")
     implementation project(":libs:tracing")
