+++ components/ledger/ledger-verification/build.gradle
@@ -42,6 +42,7 @@ dependencies {
     implementation project(':libs:messaging:messaging')
     implementation project(':libs:metrics')
     implementation project(":libs:serialization:json-validator")
+    implementation project(":libs:serialization:json-validator-common")
     implementation project(":libs:serialization:serialization-avro")
     implementation project(':libs:tracing')
     implementation project(':libs:utilities')
