+++ libs/ledger/ledger-common-data/build.gradle
@@ -24,6 +24,7 @@ dependencies {
     implementation project(':libs:flows:flow-api')
     implementation project(':libs:sandbox-types')
     implementation project(':libs:serialization:json-validator')
+    implementation project(':libs:serialization:json-validator-common')
     implementation project(':libs:serialization:serialization-amqp-api')
     implementation project(':libs:serialization:serialization-internal')
     implementation project(':libs:serialization:serialization-checkpoint-api')
