+++ libs/ledger/ledger-common-data/src/test/kotlin/net/corda/ledger/common/data/transaction/filtered/impl/FilteredTransactionImplTest.kt
@@ -3,7 +3,7 @@ package net.corda.ledger.common.data.transaction.filtered.impl
 import net.corda.cipher.suite.impl.CipherSchemeMetadataImpl
 import net.corda.cipher.suite.impl.DigestServiceImpl
 import net.corda.cipher.suite.impl.PlatformDigestServiceImpl
-import net.corda.common.json.validation.JsonValidator
+import net.corda.libs.json.validator.JsonValidator
 import net.corda.crypto.core.SecureHashImpl
 import net.corda.crypto.merkle.impl.IndexedMerkleLeafImpl
 import net.corda.crypto.merkle.impl.MerkleTreeProviderImpl
