+++ libs/serialization/json-validator-common/src/main/java/net/corda/libs/json/validator/package-info.java
@@ -1,0 +1,4 @@
+@Export
+package net.corda.libs.json.validator;
+
+import org.osgi.annotation.bundle.Export;
