+++ libs/serialization/json-validator/build.gradle
@@ -19,6 +19,7 @@ dependencies {
 
     implementation platform("net.corda:corda-api:$cordaApiVersion")
     implementation project(':libs:sandbox-types')
+    implementation project(':libs:serialization:json-validator-common')
     implementation 'net.corda:corda-base'
     implementation 'net.corda:corda-serialization'
     implementation 'org.jetbrains.kotlin:kotlin-osgi-bundle'
