+++ settings.gradle
@@ -347,6 +347,7 @@ include 'libs:schema-registry:schema-registry-impl'
 include 'libs:serialization:serialization-checkpoint-api'
 include 'libs:serialization:json-serializers'
 include 'libs:serialization:json-validator'
+include 'libs:serialization:json-validator-common'
 include 'libs:serialization:kryo-serializers'
 include 'libs:serialization:serialization-amqp'
 include 'libs:serialization:serialization-amqp-api'
