+++ testing/ledger/ledger-common-testkit/build.gradle
@@ -17,5 +17,6 @@ dependencies {
     api project(':libs:application:application-impl')
     api project(':libs:platform-info')
     api project(":libs:serialization:json-validator")
+    api project(":libs:serialization:json-validator-common")
     runtimeOnly project(':testing:ledger:ledger-hsqldb')
 }
