+++ testing/ledger/ledger-consensual-testkit/build.gradle
@@ -12,6 +12,7 @@ dependencies {
 
     implementation project(':libs:crypto:cipher-suite')
     implementation project(":libs:serialization:json-validator")
+    implementation project(":libs:serialization:json-validator-common")
     implementation project(':libs:ledger:ledger-common-data')
     implementation project(':libs:ledger:ledger-consensual-data')
     implementation project(':testing:ledger:ledger-common-testkit')
